requirements.txt
python-telegram-bot>=20.7
APScheduler>=3.10.4
python-dateutil>=2.8.2
pytz>=2024.1

